operator: move label templating functions to separate file
no code changes, just label templating functions moved to labeltmpl.go.
